@@ -1,31 +1,48 @@
 {
   "name": "babel-plugin-un-cjs",
-  "version": "1.3.0",
+  "version": "2.0.0",
   "description": "",
-  "main": "dist/index.js",
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "exports": {
+    ".": {
+      "require": "./dist/index.js",
+      "import": "./dist/index.mjs"
+    }
+  },
+  "types": "./dist/index.d.ts",
   "files": [
     "dist"
   ],
   "scripts": {
-    "start": "tsc -w",
-    "build": "tsc",
+    "build": "rollup -c rollup.config.js",
+    "build:watch": "rollup --watch -c rollup.config.js",
+    "type": "tsc",
+    "type:watch": "tsc --watch",
     "test": "jest",
+    "format": "prettier --write .",
     "test:watch": "jest --watch"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.6.4",
-    "@babel/preset-env": "^7.6.3",
-    "@babel/preset-typescript": "^7.6.0",
-    "@types/babel__core": "^7.1.3",
-    "@types/jest": "^24.0.19",
-    "@types/node": "^12.11.7",
-    "babel-jest": "^24.9.0",
-    "jest": "^24.9.0",
-    "prettier": "^1.18.2",
-    "typescript": "^3.6.4"
+    "@babel/core": "7.9.6",
+    "@rollup/plugin-node-resolve": "7.1.3",
+    "@rollup/plugin-sucrase": "3.0.1",
+    "@sucrase/jest-plugin": "2.0.0",
+    "@types/babel__core": "7.1.7",
+    "@types/babel__traverse": "7.0.9",
+    "@types/jest": "25.2.1",
+    "@types/node": "13.13.5",
+    "jest": "26.0.1",
+    "jest-circus": "26.0.1",
+    "jest-watch-typeahead": "0.6.0",
+    "prettier": "2.0.5",
+    "rollup": "2.9.1",
+    "rollup-plugin-dts": "1.4.2",
+    "source-map": "0.7.3",
+    "typescript": "3.9.2"
   },
   "prettier": {
     "semi": false,
@@ -33,6 +50,6 @@
     "trailingComma": "all"
   },
   "dependencies": {
-    "@babel/helper-plugin-utils": "^7.0.0"
+    "@babel/helper-plugin-utils": "^7.8.3"
   }
 }
