@@ -1,7 +1,7 @@
 import { NodePath, types as t } from '@babel/core'
 import { ModulePathsToReplace, NamedExportsMap } from '..'
 import { Binding } from '@babel/traverse'
-import { assignMaps, isModuleExports } from '../helpers'
+import { assignMaps, isModuleExports, isValidIdentiferName } from '../helpers'
 
 /**
  * Collects all property assignments to create a map of named exports
@@ -10,7 +10,7 @@ const getNamedExportsFromReferencedAssignments = (
   binding: Binding,
 ): NamedExportsMap => {
   const exportsMap: NamedExportsMap = new Map()
-  binding.referencePaths.forEach(refPath => {
+  binding.referencePaths.forEach((refPath) => {
     // idk if this will ever be false... but type narrowing
     if (!refPath.isIdentifier()) return
     const memExp = refPath.parentPath
@@ -35,13 +35,16 @@ const getNamedExportsFromObject = (
   obj: NodePath<t.ObjectExpression>,
 ): NamedExportsMap => {
   const exportsMap: NamedExportsMap = new Map()
-  obj.get('properties').forEach(prop => {
-    if (
-      prop.isObjectProperty() &&
-      t.isIdentifier(prop.node.key) &&
-      prop.get('value').isExpression()
-    )
-      exportsMap.set(prop.node.key.name, prop)
+  obj.get('properties').forEach((prop) => {
+    if (!prop.isObjectProperty() || !prop.get('value').isExpression()) return
+    const name = t.isIdentifier(prop.node.key)
+      ? prop.node.key.name
+      : t.isStringLiteral(prop.node.key)
+      ? prop.node.key.value
+      : undefined
+    if (!name) return
+    if (!isValidIdentiferName(name)) return
+    exportsMap.set(name, prop)
   })
   return exportsMap
 }
@@ -91,7 +94,8 @@ const getExportInfo = (path: NodePath): false | ExportInfo => {
       return {
         ...objExportInfo,
         // if it is an identifier, use it, otherwise, it is unknown
-        assignmentTarget: prop.isIdentifier() ? prop : 'unknown',
+        assignmentTarget:
+          prop.isIdentifier() && !path.node.computed ? prop : 'unknown',
       }
     }
     if (objExportInfo.assignmentTarget === 'unknown') return objExportInfo
@@ -105,17 +109,17 @@ const getExportInfo = (path: NodePath): false | ExportInfo => {
   return false
 }
 
-export const handlePotentialExport = (
+export const handleAssignmentExpression = (
   path: NodePath<t.AssignmentExpression>,
   modulePathsToReplace: ModulePathsToReplace,
   namedExports: NamedExportsMap,
 ) => {
   const left = path.get('left')
   const right = path.get('right')
-  // eliminate destructuring, nested assignments, etc.
+  // eliminate destructuring, etc.
   if (!(left.isIdentifier() || left.isMemberExpression())) return
 
-  // handle module.exports = exports.default (needs to be removed)
+  // handle module.exports = exports.default (needs to be removed, it is itself a CJS-ESM shim)
   // generated by https://github.com/59naga/babel-plugin-add-module-exports - breaks stuff
   if (
     left.isMemberExpression() &&
@@ -125,6 +129,34 @@ export const handlePotentialExport = (
     right.node.object.name === 'exports' &&
     t.isIdentifier(right.node.property) &&
     right.node.property.name === 'default'
+  )
+    return path.remove()
+
+  // handle module.exports.default = exports.default (needs to be removed)
+  // Also generated by https://github.com/59naga/babel-plugin-add-module-exports, with addDefaultProperty: true
+  if (
+    left.isMemberExpression() &&
+    t.isMemberExpression(left.node.object) &&
+    isModuleExports(left.node.object) &&
+    t.isIdentifier(left.node.property) &&
+    left.node.property.name === 'default' &&
+    right.isMemberExpression() &&
+    t.isIdentifier(right.node.object) &&
+    right.node.object.name === 'exports' &&
+    t.isIdentifier(right.node.property) &&
+    right.node.property.name === 'default'
+  )
+    return path.remove()
+
+  // Remove exports.__esModule = true
+  // Generated by babel commonjs with loose: true
+  if (
+    left.isMemberExpression() &&
+    right.isBooleanLiteral() &&
+    t.isIdentifier(left.node.object) &&
+    left.node.object.name === 'exports' &&
+    left.node.property.name === '__esModule' &&
+    right.node.value === true
   )
     return path.remove()
 
@@ -162,13 +194,52 @@ export const handlePotentialExport = (
         getNamedExportsFromReferencedAssignments(binding),
       )
     } else if (right.isObjectExpression()) {
-      // assignment: module.exports.foo = {asdf: 'asdf'}
+      // assignment: module.exports = {asdf: 'asdf'}
       // grab all properties and make them named exports
       assignMaps(namedExports, getNamedExportsFromObject(right))
+    } else {
+      // assigning module.exports = ...
+      // and the right side is not an identifier or an object
+      namedExports.set('default', path)
     }
   } else {
     // assignment: module.exports.foo = blah
-    const id = exportInfo.assignmentTarget
-    namedExports.set(id.node.name, path)
+    // or: exports.foo = blah
+    const exportName: string = exportInfo.assignmentTarget.node.name
+    // Check if there is already a module.exports.foo by the same name
+    const existingExport = namedExports.get(exportName)
+
+    // If the already-existing assignment is known to execute before this one
+    // We will determine if it can be removed
+    if (
+      existingExport &&
+      existingExport.isAssignmentExpression() &&
+      !path.willIMaybeExecuteBefore(existingExport) &&
+      evaluatesToUndefined(existingExport.node.right)
+    ) {
+      if (existingExport.parentPath.isExpressionStatement()) {
+        // It is safe to remove it altogether
+        existingExport.parentPath.remove()
+      } else {
+        // replace the assignment with just the RHS, since the RHS may have side effects
+        existingExport.replaceWith(existingExport.get('right'))
+      }
+    }
+    namedExports.set(exportName, path)
   }
+}
+
+/**
+ * Determines if a node will definitely evaluate to undefined.
+ * Only supports a limited number of use cases, otherwise returns false
+ */
+const evaluatesToUndefined = (node: t.Node): boolean => {
+  if (t.isIdentifier(node) && node.name === 'undefined') {
+    return true
+  } else if (t.isUnaryExpression(node) && node.operator === 'void') {
+    return true
+  } else if (t.isAssignmentExpression(node)) {
+    return evaluatesToUndefined(node.right)
+  }
+  return false
 }
