@@ -2,8 +2,10 @@ import { types as t, NodePath } from '@babel/core'
 import {
   getRequirePath,
   generateIdentifier,
-  getProgramBody,
-  pathsToRemove,
+  injectImportIntoBody,
+  updateReferencesTo,
+  getProgramPath,
+  importPathNameToIdentifierName,
 } from '../helpers'
 
 export const handleWildcardImport = (path: NodePath<t.CallExpression>) => {
@@ -12,32 +14,35 @@ export const handleWildcardImport = (path: NodePath<t.CallExpression>) => {
   const requireStatement = node.arguments[0]
   const importPath = getRequirePath(requireStatement)
   if (!importPath) return
-  const { parent: variableDeclarator, parentPath: varPath } = path
+  const variableDeclarator = path.parentPath
   if (
-    !t.isVariableDeclarator(variableDeclarator) ||
-    !t.isIdentifier(variableDeclarator.id)
-  )
+    !variableDeclarator.isVariableDeclarator() ||
+    !t.isIdentifier(variableDeclarator.node.id)
+  ) {
+    const newId = generateIdentifier(
+      path.scope,
+      importPathNameToIdentifierName(importPath.value),
+    )
+    const newImport = t.importDeclaration(
+      [t.importNamespaceSpecifier(newId)],
+      importPath,
+    )
+    injectImportIntoBody(getProgramPath(path), newImport)
+    path.replaceWith(newId)
     return
-  const oldId = variableDeclarator.id
-  pathsToRemove.add(varPath)
-  const varScope = path.scope
-  const globalScope = varScope.getProgramParent()
-  const globalBinding = globalScope.getBinding(oldId.name)
-  // because we are moving the variable to the global scope, it may conflict
-  // it is safe to use the original name if the declaration of the variable is in the global scope already and it is const
-  const useOriginalName =
-    globalBinding && globalBinding.path === varPath && globalBinding.constant
-  const newId = useOriginalName
-    ? oldId
-    : generateIdentifier(globalScope, oldId.name)
-  if (!useOriginalName) {
-    varScope.rename(oldId.name, newId.name)
-    globalScope.removeBinding(oldId.name)
   }
-  const importNodePath: NodePath<t.ImportDeclaration> = getProgramBody(
-    path,
-  ).insertBefore([
-    t.importDeclaration([t.importNamespaceSpecifier(newId)], importPath),
-  ])[0]
-  globalScope.registerDeclaration(importNodePath)
+  const originalId = variableDeclarator.node.id
+  const originalBinding = variableDeclarator.scope.getBinding(originalId.name)
+  if (!originalBinding) return
+  const references = originalBinding.referencePaths
+  variableDeclarator.scope.removeBinding(originalId.name)
+  const newId = generateIdentifier(variableDeclarator.scope, originalId)
+  variableDeclarator.remove()
+
+  const newImport = t.importDeclaration(
+    [t.importNamespaceSpecifier(newId)],
+    importPath,
+  )
+  injectImportIntoBody(getProgramPath(path), newImport)
+  updateReferencesTo(references, newId)
 }
