@@ -1,4 +1,4 @@
-import transform from '../test-util'
+import transform from '../../test-util'
 
 // based on https://unpkg.com/browse/@babel/types@7.6.3/lib/builders/generated/index.js
 
@@ -22,26 +22,17 @@ function Super(...args) {
 test('babel-types', async () => {
   const transformed = await transform(input)
   expect(transformed).toMatchInlineSnapshot(`
-    "\\"use strict\\";
-
-    let _default = {};
-    import _builder from \\"../builder\\";
-
-    const _super = _default.Super = Super;
-
-    export { _super as super };
-    _default.super = _super;
-
-    function _interopRequireDefault(obj) {
-      return obj && obj.__esModule ? obj : {
-        default: obj
-      };
-    }
+    "let _exports = {}
+    import _builder from '../builder'
+    export { Super as super }
+    export { Super }
+    _exports.super = _exports.Super = Super
 
     function Super(...args) {
-      return (0, _builder)(\\"Super\\", ...args);
+      return _builder('Super', ...args)
     }
 
-    export default _default;"
+    export default _exports
+    "
   `)
 })
