@@ -1,4 +1,4 @@
-import transform from '../test-util'
+import transform from '../../test-util'
 
 // based on https://unpkg.com/@babel/core@7.6.4/lib/index.js
 
@@ -58,89 +58,35 @@ function Plugin(alias) {
 test('babel', async () => {
   const transformed = await transform(input)
   expect(transformed).toMatchInlineSnapshot(`
-    "let _default = {};
-    import _template from \\"@babel/template\\";
-    import { version } from \\"../package.json\\";
-    export { Plugin };
-    _default.Plugin = Plugin;
-    export { version };
-    _default.version = version;
-    export { _template as template };
-    _default.template = _template;
-
-    function _getRequireWildcardCache() {
-      if (typeof WeakMap !== \\"function\\") return null;
-      var cache = new WeakMap();
-
-      _getRequireWildcardCache = function () {
-        return cache;
-      };
-
-      return cache;
-    }
-
-    function _interopRequireWildcard(obj) {
-      if (obj && obj.__esModule) {
-        return obj;
-      }
-
-      var cache = _getRequireWildcardCache();
-
-      if (cache && cache.has(obj)) {
-        return cache.get(obj);
-      }
-
-      var newObj = {};
-
-      if (obj != null) {
-        var hasPropertyDescriptor = Object.defineProperty && Object.getOwnPropertyDescriptor;
-
-        for (var key in obj) {
-          if (Object.prototype.hasOwnProperty.call(obj, key)) {
-            var desc = hasPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : null;
-
-            if (desc && (desc.get || desc.set)) {
-              Object.defineProperty(newObj, key, desc);
-            } else {
-              newObj[key] = obj[key];
-            }
-          }
-        }
-      }
-
-      newObj.default = obj;
-
-      if (cache) {
-        cache.set(obj, newObj);
-      }
-
-      return newObj;
-    }
-
-    function _interopRequireDefault(obj) {
-      return obj && obj.__esModule ? obj : {
-        default: obj
-      };
-    }
-
-    const DEFAULT_EXTENSIONS = Object.freeze([\\".js\\", \\".jsx\\", \\".es6\\", \\".es\\", \\".mjs\\"]);
-    export { DEFAULT_EXTENSIONS };
-    _default.DEFAULT_EXTENSIONS = DEFAULT_EXTENSIONS;
+    "let _exports = {}
+    import * as _package from '../package.json'
+    import _template from '@babel/template'
+    export { Plugin }
+    _exports.Plugin = Plugin
+    export const version = _package.version
+    _exports.version = _package.version
+    export { _template as template }
+    _exports.template = _template
+    const DEFAULT_EXTENSIONS = Object.freeze(['.js', '.jsx', '.es6', '.es', '.mjs'])
+    export { DEFAULT_EXTENSIONS }
+    _exports.DEFAULT_EXTENSIONS = DEFAULT_EXTENSIONS
 
     class OptionManager {
       init(opts) {
-        return (0, _config.loadOptions)(opts);
+        return _config.loadOptions(opts)
       }
-
     }
 
-    export { OptionManager };
-    _default.OptionManager = OptionManager;
+    export { OptionManager }
+    _exports.OptionManager = OptionManager
 
     function Plugin(alias) {
-      throw new Error(\`The (\${alias}) Babel 5 plugin is being run with an unsupported Babel version.\`);
+      throw new Error(
+        \`The (\${alias}) Babel 5 plugin is being run with an unsupported Babel version.\`,
+      )
     }
 
-    export default _default;"
+    export default _exports
+    "
   `)
 })
