@@ -1,4 +1,4 @@
-import transform from '../test-util'
+import transform from '../../test-util'
 
 // based on https://unpkg.com/micromatch@3.1.0/lib/parsers.js
 
@@ -22,23 +22,24 @@ module.exports.sdf = function () {
 exports.asdf = nanomatch
 `
   expect(await transform(input)).toMatchInlineSnapshot(`
-    "import _toRegex from 'to-regex';
-    import { create } from 'regex-not';
-    import _nanomatch from 'nanomatch';
-    import { parsers } from 'extglob';
+    "import * as extglob from 'extglob'
+    import nanomatch from 'nanomatch'
+    import * as regexNot from 'regex-not'
+    import toRegex from 'to-regex'
 
-    let _default = function (snapdragon) {
-      console.log('nanomatch', _nanomatch.parsers);
-      console.log('extglob', parsers);
-    };
+    let _exports = function (snapdragon) {
+      console.log('nanomatch', nanomatch.parsers)
+      console.log('extglob', extglob.parsers)
+    }
 
     export const sdf = function () {
-      create();
-      return _toRegex();
-    };
-    _default.sdf = sdf;
-    export { _nanomatch as asdf };
-    _default.asdf = _nanomatch;
-    export default _default;"
+      regexNot.create()
+      return toRegex()
+    }
+    _exports.sdf = sdf
+    export { nanomatch as asdf }
+    _exports.asdf = nanomatch
+    export default _exports
+    "
   `)
 })
