@@ -1,4 +1,4 @@
-import transform from '../test-util'
+import transform from '../../test-util'
 
 test('wait-for-expect', async () => {
   const input = `
@@ -23,21 +23,17 @@ module.exports.default = exports.default
 `
   const transformed = await transform(input)
   expect(transformed).toMatchInlineSnapshot(`
-    "'use strict';
+    "let _exports = {}
+    const defaults = {}
 
-    let _default2 = {};
-    _default2.default = void 0;
-    const defaults = {};
-
-    var waitForExpect = function waitForExpect(expectation) {};
+    var waitForExpect = function waitForExpect(expectation) {}
 
     var _default = Object.assign(waitForExpect, {
-      defaults: defaults
-    });
+      defaults: defaults,
+    })
 
-    _default2.default = _default;
-    const _default3 = _default2.default;
-    export { _default3 as default };
-    _default2.default = _default3;"
+    export default _default
+    _exports.default = _default
+    "
   `)
 })
