@@ -1,22 +1,4 @@
-import transform from './test-util'
-
-test('replaces global with window', async () => {
-  const input = `
-console.log(global.asdf)
-`
-  const transformed = await transform(input)
-  expect(transformed).toMatchInlineSnapshot(`"console.log(window.asdf);"`)
-})
-
-test('removes __esModule', async () => {
-  const input = `
- Object.defineProperty(exports, '__esModule', {
-  value: true
-}); 
-`
-  const transformed = await transform(input)
-  expect(transformed).toMatchInlineSnapshot(`""`)
-})
+import transform from '../test-util'
 
 describe('handle export object', () => {
   test('handles Object.defineProperty on exports', async () => {
@@ -31,11 +13,12 @@ Object.defineProperty(exports, "version", {
   `
     const transformed = await transform(input)
     expect(transformed).toMatchInlineSnapshot(`
-      "let _default = {};
-      import { version } from \\"../package.json\\";
-      export { version };
-      _default.version = version;
-      export default _default;"
+      "let _exports = {}
+      import * as _package from '../package.json'
+      export const version = _package.version
+      _exports.version = _package.version
+      export default _exports
+      "
     `)
   })
   describe('handles objects', () => {
@@ -50,18 +33,19 @@ module.exports.foobarbaz = 'foobarbaz'
 `
       const transformed = await transform(input)
       expect(transformed).toMatchInlineSnapshot(`
-        "const foo = 'foo';
-        _default.foobar = 'foobar';
+        "const foo = 'foo'
+        _exports.foobar = 'foobar'
 
-        const _foo = () => {};
+        const _foo = () => {}
 
-        export { _foo as foo };
-        let _default = {
-          foo: _foo
-        };
-        export const foobarbaz = 'foobarbaz';
-        _default.foobarbaz = foobarbaz;
-        export default _default;"
+        export { _foo as foo }
+        let _exports = {
+          foo: _foo,
+        }
+        export const foobarbaz = 'foobarbaz'
+        _exports.foobarbaz = foobarbaz
+        export default _exports
+        "
       `)
     })
     test('export property on referenced object', async () => {
@@ -75,16 +59,17 @@ module.exports.otherExport = 'otherExport'
 `
       const transformed = await transform(input)
       expect(transformed).toMatchInlineSnapshot(`
-        "export const foo = () => {};
+        "export const foo = () => {}
         const obj = {
-          foo: foo
-        };
-        let _default = obj;
-        export const foobarbaz = 'foobarbaz';
-        obj.foobarbaz = foobarbaz;
-        export const otherExport = 'otherExport';
-        _default.otherExport = otherExport;
-        export default _default;"
+          foo: foo,
+        }
+        let _exports = obj
+        export const foobarbaz = 'foobarbaz'
+        obj.foobarbaz = foobarbaz
+        export const otherExport = 'otherExport'
+        _exports.otherExport = otherExport
+        export default _exports
+        "
       `)
     })
   })
@@ -115,12 +100,13 @@ module.exports = main
 `
       const transformed = await transform(input)
       expect(transformed).toMatchInlineSnapshot(`
-        "let main = () => {};
+        "let main = () => {}
 
-        export const other = 'hiiiii';
-        main.other = other;
-        let _default = main;
-        export default _default;"
+        export const other = 'hiiiii'
+        main.other = other
+        let _exports = main
+        export default _exports
+        "
       `)
     })
   })
@@ -144,9 +130,8 @@ Object.keys(_waitFor).forEach(function (key) {
 `
   const transformed = await transform(input)
   expect(transformed).toMatchInlineSnapshot(`
-    "let _default = {};
-    export * from \\"./wait-for\\";
-    export default _default;"
+    "export * from './wait-for'
+    "
   `)
 })
 
@@ -167,9 +152,8 @@ Object.keys(_waitFor).forEach(function (key) {
 `
   const transformed = await transform(input)
   expect(transformed).toMatchInlineSnapshot(`
-    "let _default = {};
-    export * from \\"./wait-for\\";
-    export default _default;"
+    "export * from './wait-for'
+    "
   `)
 })
 
@@ -201,9 +185,10 @@ test('destructured require', async () => {
 const { docsUrl, foo: bar, default: utils } = require('../utilities');
 `
   const transformed = await transform(input)
-  expect(transformed).toMatchInlineSnapshot(
-    `"import { docsUrl, foo as bar, default as utils } from '../utilities';"`,
-  )
+  expect(transformed).toMatchInlineSnapshot(`
+    "import { docsUrl, foo as bar, default as utils } from '../utilities'
+    "
+  `)
 })
 
 describe('uses single declaration for _default if possible', () => {
@@ -216,12 +201,13 @@ module.exports = {
 `
     const transformed = await transform(input)
     expect(transformed).toMatchInlineSnapshot(`
-      "_default.bar = 'hi';
-      export const foo = () => {};
-      let _default = {
-        foo: foo
-      };
-      export default _default;"
+      "_exports.bar = 'hi'
+      export const foo = () => {}
+      let _exports = {
+        foo: foo,
+      }
+      export default _exports
+      "
     `)
   })
 
@@ -237,13 +223,14 @@ exports = {
     // And in the above module.exports code foo is exported but not here
     // It is because assigning to the `exports` object does not clear out the `module.exports` object
     expect(transformed).toMatchInlineSnapshot(`
-      "let _default = {};
-      export const bar = 'hi';
-      _default.bar = bar;
-      _default = {
-        foo: () => {}
-      };
-      export default _default;"
+      "let _exports = {}
+      export const bar = 'hi'
+      _exports.bar = bar
+      _exports = {
+        foo: () => {},
+      }
+      export default _exports
+      "
     `)
   })
 })
